requirements.txt
google-generativeai
openai==1.20.0
pandas==2.2.2
protobuf==4.25.3
pydantic==2.7.0
Pympler==1.0.1
click==8.1.7
ollama
fastapi
uvicorn
anthropic
python-dotenv
platformdirs
watchdog
